+++ app/dashboard/page.jsx
@@ -1,12 +1,13 @@
 import React from "react";
 import HomePage from "@/app/HomePage/page";
+import { buildMetadata } from "@/lib/seo";
 
-function Dashboard() {
+export const metadata = buildMetadata("/dashboard");
+
+export default function DashboardPage() {
   return (
     <div className="">
       <HomePage />
     </div>
   );
 }
-
-export default Dashboard;
