+++ app/dashboard/radioserverhosting/page.jsx
@@ -1,213 +1,8 @@
-"use client";
+import RadioServerHostingClient from "./RadioServerHostingClient";
+import { buildMetadata } from "@/lib/seo";
 
-import Link from "next/link";
-import React from "react";
-import { PiArrowLeftThin } from "react-icons/pi";
-import { motion } from "framer-motion";
-import Image from "next/image";
+export const metadata = buildMetadata("/dashboard/radioserverhosting");
 
-function Page() {
-  return (
-    <motion.div
-      className="text-neutral-50"
-      initial={{ y: 100, opacity: 0 }}
-      animate={{
-        y: 0,
-        opacity: 1,
-        transition: {
-          duration: 0.8,
-          delay: 0.9,
-          type: "spring",
-          stiffness: 200,
-        },
-      }}
-    >
-      <div className="sticky top-5">
-        <div>
-          <div className="-mt-6">
-            <div className="bg-neutral-700/25 backdrop-blur-md h-10 w-full rounded-xl flex items-center gap-x-7">
-              <Link href={"/dashboard"}>
-                <div className="bg-neutral-700/50 h-8 w-8 rounded-full flex items-center justify-center ml-3">
-                  <div className="bg-neutral-300 rounded-full h-5 w-5 flex items-center justify-center">
-                    <PiArrowLeftThin className="text-black text-lg" />
-                  </div>
-                </div>
-              </Link>
-              <Link href={"/dashboard"}>
-                <button className="text-xs bg-neutral-700/25 p-1 w-16 h-6 rounded-md">
-                  Kezdőlap
-                </button>
-              </Link>
-              <Link href={"/dashboard/hostingbuilder"}>
-                <button className="text-xs bg-neutral-700/25 p-1 w-16 h-6 rounded-md">
-                  Hoszting
-                </button>
-              </Link>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="mt-9 w-full p-5 border border-neutral-700 rounded-2xl h-full bg-[#1C1C1C]">
-        <motion.h1
-          initial={{ x: 100, opacity: 0, filter: "blur(4px)" }}
-          animate={{
-            x: 0,
-            opacity: 1,
-            filter: "blur(0px)",
-            transition: {
-              duration: 0.8,
-              delay: 0.9,
-              type: "spring",
-              stiffness: 200,
-            },
-          }}
-          transition={{ delay: 0.4 }}
-          className="text-3xl font-RubikExtraBold text-center"
-        >
-          Rádiószerver hoszting szolgáltatás kiépitése
-        </motion.h1>
-        <div className="flex items-center gap-x-2 justify-center my-4 font-RubikRegular">
-          <p className="bg-[#282828] w-fit text-neutral-300 rounded-md px-2 h-5 flex items-center justify-center text-[10px]">
-            rádiószerver, azuracast
-          </p>
-          <div className="w-1 h-1 rounded-full bg-neutral-400" />
-          <span className="text-xs">Frissítve: 2024.07.09.</span>
-        </div>
-        <div className="my-11 font-RubikMedium">
-          <Image
-            width={1000}
-            height={1000}
-            className="rounded-lg h-56 w-full object-cover"
-            src="/internet-radio-station.webp"
-            alt="internet-radio-station"
-          />
-          <p className="my-7">
-            A cikk fő részeiben bemutatom a legfontosabb szolgáltatásokat, amik
-            ide kapcsolódnak:
-          </p>
-          <ul>
-            <li>
-              <strong>Webhoszting:</strong> Ez föleg hosztingod saját
-              weboldalának tárolására szolgál.
-            </li>
-            <li>
-              <strong>E-mail hoszting:</strong> Az e-mail küldésére szolgál, ez
-              többnyire tartalmazza a webhoszting. Ügyfelek kiértesitése
-              szolgáltatásaikról, azok állapotáról.
-            </li>
-            <li>
-              <strong>VPS hoszting:</strong> Virtuális privát szerverek
-              biztosítása. Ez nagyban függ, hogyan szeretnéd a rádiószervereket
-              biztositani.
-            </li>
-          </ul>
-          <div className="my-9">
-            <Image
-              height={1000}
-              width={1000}
-              className="rounded-lg h-56 w-full object-cover"
-              src="/radioserverlive.gif"
-              alt="radioserverlive"
-            />
-          </div>
-          <h2 className="text-2xl font-RubikExtraBold">
-            Hogyan történik a Rádiószerver Hosztolás?{" "}
-          </h2>
-          <p className="my-4">
-            A rádiószerver hosztolás egyre népszerűbbé válik, köszönhetően az
-            olyan könnyen használható platformoknak, mint az Azuracast. Ez a
-            folyamat nemcsak a megfelelő szerverek és szoftverek telepítését
-            jelenti, hanem jogi és technikai lépések sorozatát is, amelyek
-            biztosítják a rádióállomás zavartalan működését és jogszerűségét.
-            Ebben a cikkben áttekintjük a rádiószerver hosztolás lépéseit és a
-            fontos szempontokat, amelyekre figyelni kell.
-          </p>
-          <h2 className="text-2xl font-RubikBold my-5">
-            1. Ügyfél igényeinek és elképzeléseinek megértése{" "}
-          </h2>
-          <p className="my-4">
-            A rádiószerver hosztolás megkezdése előtt fontos megérteni az ügyfél
-            igényeit és elképzeléseit. Milyen típusú rádióadást kívánnak
-            sugározni? Mekkora hallgatói bázisra számítanak? Milyen speciális
-            funkciókat kell támogatnia a rendszernek? Ezek a kérdések segítenek
-            meghatározni a megfelelő technikai megoldásokat.
-          </p>
-          <h2 className="text-2xl font-RubikBold my-5">
-            2. Szükséges infrastruktúra kiválasztása{" "}
-          </h2>
-          <p className="my-4">
-            Az igények felmérése után a szolgáltató kiválasztja a szükséges
-            infrastruktúrát. Ez lehet fizikai szerver, virtuális privát szerver
-            (VPS) vagy felhőalapú megoldás. Az Azuracast például egyszerűen
-            telepíthető és konfigurálható különböző típusú szervereken, így
-            rugalmasan alkalmazkodik az ügyfél igényeihez.
-          </p>
-          <h2 className="text-2xl font-RubikBold my-5">
-            3. Hardver és szoftver konfiguráció
-          </h2>
-          <p className="my-4">
-            Miután kiválasztották az infrastruktúrát, a szolgáltató beállítja és
-            konfigurálja a szükséges hardveres és szoftveres elemeket. Ez
-            magában foglalja a szerverek telepítését, az Azuracast vagy hasonló
-            rádiószerver szoftver telepítését és konfigurálását, valamint a
-            biztonsági beállítások és a tűzfalak konfigurálását.
-          </p>
-          <h2 className="text-2xl font-RubikBold my-5">
-            4. Rádiószerver szolgáltatás beállítása
-          </h2>
-          <p className="my-4">
-            A következő lépésben a rádiószerver szolgáltatás kerül beállításra.
-            Ez magában foglalja a rádióállomás létrehozását az Azuracast
-            platformon, a zenei és egyéb tartalmak feltöltését, valamint a
-            műsorok és adások ütemezését. Fontos, hogy a hallgatók számára
-            biztosítsák a megfelelő minőségű streamelést.
-          </p>
-          <h2 className="text-2xl font-RubikBold my-5">
-            5. Jogi követelmények és jogdíjak kezelése
-          </h2>
-          <p className="my-4">
-            A rádiószerver hosztolás során kiemelten fontos a jogi követelmények
-            betartása. Ez magában foglalja a szerzői jogi kérdéseket, a zenei
-            tartalmak licencelését és a jogdíjak fizetését. Az ügyfélnek
-            biztosítania kell, hogy minden sugárzott tartalom jogszerűen
-            kerüljön adásba, és rendelkezzenek a szükséges engedélyekkel.
-          </p>
-          <h2 className="text-2xl font-RubikBold my-5">
-            6. Biztonsági intézkedések és mentési stratégiák beállítása
-          </h2>
-          <p className="my-4">
-            A biztonság kulcsfontosságú a rádiószerver hosztolás során is. A
-            szolgáltatóknak gondoskodniuk kell az adások és a hallgatók
-            adatainak védelméről, SSL tanúsítványok telepítésével, rendszeres
-            biztonsági frissítések végrehajtásával és a tűzfalak megfelelő
-            beállításával. Emellett fontos a rendszeres mentések készítése is.
-          </p>
-          <h2 className="text-2xl font-RubikBold my-5">
-            7. Ügyfél kapcsolattartás és támogatás
-          </h2>
-          <p className="my-4">
-            A rádiószerver hosztolás folyamatos ügyféltámogatást igényel. A
-            szolgáltatóknak biztosítaniuk kell a gyors és hatékony
-            problémamegoldást, az ügyfélszolgálat elérhetőségét, és szükség
-            esetén a technikai segítségnyújtást. Az ügyfelekkel való jó
-            kommunikáció és támogatás kulcsfontosságú a sikeres rádiószerver
-            hosztolásban.
-          </p>
-          <p className="my-4">
-            Összegzés: A rádiószerver hosztolás egy összetett folyamat, amelyet
-            alapos tervezés, konfiguráció és jogi megfelelőség előz meg. Az
-            Azuracast és hasonló platformok megkönnyítik a rádiószerverek
-            telepítését és üzemeltetését, de a megfelelő infrastruktúra
-            kiválasztása, a biztonsági intézkedések és a folyamatos
-            ügyféltámogatás elengedhetetlenek a sikeres működéshez. A szakértői
-            tudás és a tapasztalat nélkülözhetetlen a jogi és technikai
-            kihívások megoldásában, valamint a megbízható rádiósugárzás
-            biztosításában.
-          </p>
-        </div>
-      </div>
-    </motion.div>
-  );
+export default function RadioServerHostingPage() {
+  return <RadioServerHostingClient />;
 }
-
-export default Page;
