+++ app/layout.js
@@ -1,16 +1,29 @@
 import "./globals.css";
+import { Rubik } from "next/font/google";
 import Theming from "@/components/providers/Theme";
+import { ConsentProvider } from "@/components/providers/ConsentProvider";
+import AnalyticsGate from "@/components/providers/AnalyticsGate";
+import ConsentBanner from "@/components/consent/ConsentBanner";
+import SiteFooter from "@/components/layout/SiteFooter";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
+import { baseUrl, siteMetadata } from "@/lib/seo";
 
 export const runtime = "edge";
+
+const rubik = Rubik({
+  subsets: ["latin", "latin-ext"],
+  display: "swap",
+  variable: "--font-rubik",
+  weight: ["300", "400", "500", "700"],
+});
 
 const structuredData = {
   "@context": "https://schema.org",
   "@type": ["LocalBusiness", "ProfessionalService"],
   name: "PromNET - Polyák Csaba E.V.",
-  image: "https://www.promnet.hu/logo-white.png",
-  url: "https://www.promnet.hu",
+  image: `${baseUrl}/logo-white.png`,
+  url: baseUrl,
   telephone: "+36 20 549 4107",
   email: "info@promnet.hu",
   description:
@@ -23,7 +36,7 @@ const structuredData = {
     addressCountry: "HU",
   },
   sameAs: [
-    "https://www.promnet.hu",
+    baseUrl,
     "https://github.com/promnet",
     "https://www.linkedin.com/in/csaba-polyak-3497b0133/",
   ],
@@ -39,33 +52,78 @@ const structuredData = {
   ],
 };
 
+export const metadata = {
+  metadataBase: new URL(baseUrl),
+  title: {
+    default: siteMetadata.title,
+    template: "%s | PromNET",
+  },
+  description: siteMetadata.description,
+  keywords: siteMetadata.keywords,
+  openGraph: {
+    title: siteMetadata.title,
+    description: siteMetadata.description,
+    url: baseUrl,
+    siteName: "PromNET",
+    locale: "hu_HU",
+    type: "website",
+    images: [
+      {
+        url: `${baseUrl}/logo-white.png`,
+        width: 1200,
+        height: 630,
+        alt: "PromNET logó",
+      },
+    ],
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: siteMetadata.title,
+    description: siteMetadata.description,
+    images: [`${baseUrl}/logo-white.png`],
+  },
+  alternates: {
+    canonical: baseUrl,
+  },
+  icons: {
+    icon: [
+      { url: "/favicon.svg", type: "image/svg+xml" },
+    ],
+    apple: "/favicon.svg",
+  },
+  robots: {
+    index: true,
+    follow: true,
+    googleBot: {
+      index: true,
+      follow: true,
+    },
+  },
+};
+
 export default function RootLayout({ children }) {
   return (
-    <html lang="hu">
+    <html lang="hu" suppressHydrationWarning>
       <head>
-        <title>PromNET - Polyák Csaba webfejlesztés és üzemeltetés</title>
-        <meta
-          name="description"
-          content="Polyák Csaba (PromNET) egyedi weboldalakat, hosting megoldásokat és informatikai szerviz szolgáltatásokat kínál magyar vállalkozásoknak."
-        />
-        <meta property="og:title" content="PromNET - Polyák Csaba" />
-        <meta
-          property="og:description"
-          content="Professzionális webfejlesztés, digitális termék- és üzemeltetési szolgáltatások KKV-knak és alkotóknak Magyarországon."
-        />
-        <meta property="og:image" content="/logo-white.png" />
-        <meta property="og:url" content="https://www.promnet.hu" />
-        <meta name="twitter:card" content="summary_large_image" />
-        <link rel="canonical" href="https://www.promnet.hu" />
         <script
           type="application/ld+json"
           dangerouslySetInnerHTML={{ __html: JSON.stringify(structuredData) }}
         />
       </head>
-      <body className="bg-[#171717]">
-        <Theming>{children}</Theming>
-        <Analytics />
-        <SpeedInsights />
+      <body className={`${rubik.variable} min-h-screen bg-background text-foreground font-sans`}>
+        <Theming>
+          <ConsentProvider>
+            <div className="flex min-h-screen flex-col">
+              <main className="flex-1">{children}</main>
+              <SiteFooter />
+            </div>
+            <ConsentBanner />
+            <AnalyticsGate>
+              <Analytics />
+              <SpeedInsights />
+            </AnalyticsGate>
+          </ConsentProvider>
+        </Theming>
       </body>
     </html>
   );
