+++ app/robots.js
@@ -1,0 +1,12 @@
+import { baseUrl } from "@/lib/seo";
+
+export default function robots() {
+  return {
+    rules: {
+      userAgent: "*",
+      allow: "/",
+    },
+    sitemap: `${baseUrl}/sitemap.xml`,
+    host: baseUrl,
+  };
+}
