+++ app/sitemap.js
@@ -1,0 +1,5 @@
+import { sitemapEntries } from "@/lib/seo";
+
+export default function sitemap() {
+  return sitemapEntries();
+}
