+++ lib/analytics.js
@@ -3,6 +3,10 @@ export const trackEvent = (eventName, payload = {}) => {
     return;
   }
 
+  if (window.__promnetConsent !== "granted") {
+    return;
+  }
+
   if (window.va && typeof window.va.track === "function") {
     window.va.track(eventName, payload);
   }
