+++ tailwind.config.js
@@ -12,50 +12,37 @@ module.exports = {
 
   theme: {
     extend: {
-      backgroundImage: {
-        gradient:
-          "linear-gradient(145.37deg, rgba(255, 255, 255, 0.09) -8.75%, rgba(255, 255, 255, 0.027) 83.95%)",
-        gradientHover:
-          "linear-gradient(145.37deg, rgba(255, 255, 255, 0.1) -8.75%, rgba(255, 255, 255, 0.057) 83.95%)",
-        shine:
-          "linear-gradient(45deg, rgba(255,255,255,0) 45%,rgba(255,255,255,1) 50%,rgba(255,255,255,0) 55%,rgba(255,255,255,0) 100%)",
-      },
-
-      animation: {
-        blob: "blob 7s infinite",
-      },
-
-      keyframes: {
-        blob: {
-          "0%": {
-            transform: "translate(0px, 0px) scale(1)",
-          },
-          "33%": {
-            transform: "translate(30px, -50px) scale(1.1)",
-          },
-          "66%": {
-            transform: "translate(-20px, 20px) scale(0.9)",
-          },
-          "100%": {
-            transform: "tranlate(0px, 0px) scale(1)",
-          },
+      colors: {
+        background: "var(--color-background)",
+        foreground: "var(--color-foreground)",
+        muted: "var(--color-muted)",
+        border: "var(--color-border)",
+        card: "var(--color-card)",
+        accent: {
+          DEFAULT: "var(--color-accent)",
+          foreground: "var(--color-accent-foreground)",
         },
-        shinee: {
-          "0%": { backgroundPosition: "-80%" },
-          "100%": { backgroundPosition: "80%" },
-        },
-        dash: {
-          "0%": { strokeDashoffset: 1000 },
-          "100%": { strokeDashoffset: 0 },
-        },
+        success: "var(--color-success)",
+        warning: "var(--color-warning)",
       },
-    },
-    fontFamily: {
-      RubikBold: "RubikBold",
-      RubikExtraBold: "RubikExtraBold",
-      RubikLight: "RubikLight",
-      RubikMedium: "RubikMedium",
-      RubikRegular: "RubikRegular",
+      borderRadius: {
+        xl: "1.5rem",
+      },
+      boxShadow: {
+        soft: "0 20px 60px -30px rgba(124, 58, 237, 0.4)",
+        accent: "0 25px 80px -40px rgba(236, 72, 153, 0.55)",
+      },
+      fontFamily: {
+        sans: ["var(--font-rubik)", "system-ui", "sans-serif"],
+        RubikBold: ["var(--font-rubik)", "system-ui", "sans-serif"],
+        RubikExtraBold: ["var(--font-rubik)", "system-ui", "sans-serif"],
+        RubikLight: ["var(--font-rubik)", "system-ui", "sans-serif"],
+        RubikMedium: ["var(--font-rubik)", "system-ui", "sans-serif"],
+        RubikRegular: ["var(--font-rubik)", "system-ui", "sans-serif"],
+      },
+      gridTemplateColumns: {
+        responsive: "repeat(auto-fit, minmax(160px, 1fr))",
+      },
     },
   },
